+++ anreader_sources/device/jumps/n3jump.h
@@ -55,7 +55,7 @@ public:
     std::unique_ptr<t_jump_attribute> getPairs() const override;
     virtual void setPairs(const t_jump_attribute& pairs) override;
     static const QString& field_name(const int n_field);
-    static int index(const QString& field_name);
+    static size_t index(const QString& field_name);
 
 private:   
     uint m_exit_alt;
